+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/operator/rel/RelOpIterate.kt
@@ -6,6 +6,7 @@ import org.partiql.eval.ExprValue
 import org.partiql.eval.Row
 import org.partiql.eval.internal.helpers.DatumUtils.lowerSafe
 import org.partiql.eval.internal.helpers.PErrors
+import org.partiql.eval.internal.helpers.checkInterrupted
 import org.partiql.spi.types.PType
 import org.partiql.spi.value.Datum
 
@@ -37,6 +38,7 @@ internal class RelOpIterate(
     }
 
     override fun next(): Row {
+        checkInterrupted()
         val i = index
         val v = iterator.next()
         index += 1