+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/operator/rel/RelOpJoinOuterRight.kt
@@ -5,6 +5,7 @@ import org.partiql.eval.ExprRelation
 import org.partiql.eval.ExprValue
 import org.partiql.eval.Row
 import org.partiql.eval.internal.helpers.ValueUtility.isTrue
+import org.partiql.eval.internal.helpers.checkInterrupted
 import org.partiql.plan.rel.RelType
 import org.partiql.spi.value.Datum
 
@@ -69,6 +70,7 @@ internal class RelOpJoinOuterRight(
             var rhsMatched = false
             lhs.open(env)
             for (lhsRecord in lhs) {
+                checkInterrupted()
                 val input = lhsRecord.concat(rhsRecord)
                 val result = condition.eval(env.push(input))
                 if (result.isTrue()) {