+++ partiql-eval/src/main/kotlin/org/partiql/eval/internal/operator/rel/RelOpProject.kt
@@ -4,6 +4,7 @@ import org.partiql.eval.Environment
 import org.partiql.eval.ExprRelation
 import org.partiql.eval.ExprValue
 import org.partiql.eval.Row
+import org.partiql.eval.internal.helpers.checkInterrupted
 
 internal class RelOpProject(
     private val input: ExprRelation,
@@ -22,6 +23,7 @@ internal class RelOpProject(
     }
 
     override fun next(): Row {
+        checkInterrupted()
         val r = input.next()
         val p = projections.map { it.eval(env.push(r)) }.toTypedArray()
         return Row(p)